Added code to reduce the system tick rate to the minimum, thus allowing Sleep(1) to be as close to 1MS as possible.
Added support for DIAG 0 which required to allow a SYSDUMP to be taken after pressing Reset/Run.

Added a Reset button to the virtual maintenance panel.

Added an indicator to the virtual maintenance panel showing how much data is currently in the printer output file.

Changed all .CFG files to have a 1MS I/O delay.